requirements.txt
fastapi
uvicorn
tensorflow==2.19.1
numpy
pillow
pytest
python-multipart
httpx